+++ src/commands/deploy/index.ts
@@ -8,11 +8,13 @@ import { createDeployLogsCommand } from "./logs";
 import { createDeployEditingHostCommand } from "./editing-host";
 import { createDeploySiteCommand } from "./site";
 import { createDeployBuildConfigCommand } from "./build-config";
+import { createDeployEnvFileCommand } from "./env-file";
 
 export const createDeployCommand = (): Command => {
   const command = new Command("deploy").description("XM Cloud Deploy API commands");
 
   command.addCommand(createDeployBuildConfigCommand());
+  command.addCommand(createDeployEnvFileCommand());
   command.addCommand(createDeployDeploymentsCommand());
   command.addCommand(createDeployEditingHostCommand());
   command.addCommand(createDeployEnvironmentsCommand());